\name{glmmEPvignette}
\alias{glmmEPvignette}
\title{Display the package's vignette.}
\description{The vignette of the glmmEP package is displayed using the default PDF file browser. It provides a detailed detailed description of use of the package for binary mixed model analysis.}
\usage{
glmmEPvignette()
}
\author{Matt Wand\email{matt.wand@uts.edu.au} and James Yu\email{james.yu@student.uts.edu.au}}
\examples{
\donttest{glmmEPvignette()}
}

